+++ packages/components/src/components/ScalarTextInput/index.ts
@@ -1,1 +1,2 @@
 export { default as ScalarTextInput } from './ScalarTextInput.vue'
+export { default as ScalarTextInputCopy } from './ScalarTextInputCopy.vue'