+++ packages/workspace-store/src/schemas/v3.1/strict/openapi-document.ts
@@ -12,6 +12,10 @@ import {
   type XScalarOriginalDocumentHash,
   XScalarOriginalDocumentHashSchema,
 } from '@/schemas/extensions/document/x-scalar-original-document-hash'
+import {
+  type XScalarActiveEnvironment,
+  XScalarActiveEnvironmentSchema,
+} from '@/schemas/extensions/general/x-scalar-active-environment'
 import { type XScalarCookies, xScalarCookiesSchema } from '@/schemas/extensions/general/x-scalar-cookies'
 import { type XScalarOrder, XScalarOrderSchema } from '@/schemas/extensions/general/x-scalar-order'
 import {
@@ -73,7 +77,6 @@ import { XMLObjectSchemaDefinition } from './xml'
 export const OpenApiExtensionsSchema = compose(
   Type.Partial(
     Type.Object({
-      'x-scalar-client-config-active-environment': Type.String(),
       'x-original-oas-version': Type.String(),
       'x-scalar-original-source-url': Type.String(),
       [extensions.document.navigation]: TraversedDocumentObjectRef,
@@ -87,10 +90,10 @@ export const OpenApiExtensionsSchema = compose(
   xScalarCookiesSchema,
   XScalarOriginalDocumentHashSchema,
   XScalarIsDirtySchema,
+  XScalarActiveEnvironmentSchema,
 )
 
 export type OpenAPIExtensions = Partial<{
-  'x-scalar-client-config-active-environment': string
   'x-original-oas-version': string
   /** Original document source url / when loading a document from an external source */
   'x-scalar-original-source-url': string
@@ -99,6 +102,7 @@ export type OpenAPIExtensions = Partial<{
   XScalarOriginalDocumentHash &
   XTagGroups &
   XScalarEnvironments &
+  XScalarActiveEnvironment &
   XScalarSelectedServer &
   XScalarIcon &
   XScalarOrder &