+++ packages/snippetz/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @scalar/snippetz
 
+## 0.6.17
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/types@0.6.8**
+  - [#8274](https://github.com/scalar/scalar/pull/8274): feat(agent): add config to hide search api
+
 ## 0.6.16
 
 ### Patch Changes