+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @scalar/types
 
+## 0.6.8
+
+### Patch Changes
+
+- [#8274](https://github.com/scalar/scalar/pull/8274): feat(agent): add config to hide search api
+
 ## 0.6.7
 
 ### Patch Changes